airflow_ge/requirements.txt
asn1crypto==1.5.1
boto3==1.26.83
botocore==1.29.83
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.2
filelock==3.9.0
idna==3.4
jmespath==1.0.1
numpy==1.24.2
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
snowflake-connector-python==3.0.1
typing_extensions==4.5.0
urllib3==1.26.14
